fix more things related to settings.py
